+++ meshers/blocky/voxel_blocky_model_mesh.h
@@ -25,11 +25,17 @@ public:
 	void rotate_90(math::Axis axis, bool clockwise) override;
 	void rotate_ortho(math::OrthoBasis p_ortho_basis) override;
 
+	void set_side_vertex_tolerance(float tolerance);
+	float get_side_vertex_tolerance() const;
+
 private:
 	static void _bind_methods();
 
 	Ref<Mesh> _mesh;
 	uint8_t _mesh_ortho_rotation = 0;
+	// Margin near sides of the voxel where triangles will be considered to be "on the side". Those triangles will
+	// be processed by the neighbor side culling system.
+	float _side_vertex_tolerance = 0.001f;
 };
 
 } // namespace zylann::voxel